Mobile-Agent-v3.5/browser_use/browser/playwright/Online_Mind2Web_judge/requirements.txt
backoff==2.2.1
openai==1.68.2
Pillow==11.1.0
